Make Node orchestrate processing Actions / Events
This is the first step towards cleaning up the top-level structure.
The Node now mostly deals with modules and orchestrates the flow of
Actions and events between them.

The interface for submitting client request is simplified and becomes
a single method `SubmitRequest` of the Node.
