+++ go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/stretchr/testify v1.6.1 // indirect
 	github.com/tidwall/wal v0.1.3
 	golang.org/x/sys v0.0.0-20200821140526-fda516888d29 // indirect
-	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
 	google.golang.org/protobuf v1.26.0
 	gopkg.in/alecthomas/kingpin.v2 v2.2.6
 	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
