+++ pkg/processor/msgfilter.go
@@ -15,7 +15,7 @@ import (
 // TODO, we probably can/should add some more basic error checking here.  Particularly
 // identifying pointer fields which must be set.
 
-func preProcess(outerMsg *msgs.Msg) error {
+func PreProcess(outerMsg *msgs.Msg) error {
 	switch innerMsg := outerMsg.Type.(type) {
 	case *msgs.Msg_Preprepare:
 		if innerMsg.Preprepare == nil {
